statistics page initialized
