+++ src/App.js
@@ -1,17 +1,15 @@
 import React from 'react';
-import logo from './logo.svg';
 import './App.css';
 import 'antd/dist/antd.css';
 import moment from 'moment';
 import 'moment/locale/zh-cn';
-import MapBox from "./MapBox";
-import LineChart from "./LineChart";
+import StatisticsView from "./view/StatisticsView";
 moment.locale('zh-cn');
 
 function App() {
   return (
     <div className="App">
-      <LineChart />
+        <StatisticsView />
     </div>
   );
 }
