+++ src/components/LineChart.js
@@ -1,16 +1,23 @@
 import React from "react";
 import { Line }from "@ant-design/charts";
-import {testData} from "./MapBox.js"
+/*
+LineChart:
+@author Zhuohao Shen <ao7777@sjtu.edu.cn>
+@date July 7th 2020
+@description A chart of line showing visit counts by hours
+@params {Array(JSON)} data - data of distribution order by hour
+@contents {DOM} - a flex box of chart showing distribution
+*/
 export default class LineChart extends React.Component{
     state={
         data:[],
         config:{},
     }
-    constructor() {
-        super();
-        this.state={data: testData[0].time_distr,
+    constructor(props) {
+        super(props);
+        this.state={data: props.data,
         config: {
-            data:testData[0].time_distr,
+            data:props.data,
             title:{visible:true,text:'分时段统计'},
             description: {
                 visible: true,
